requirements.txt
streamlit
groq
sentence-transformers
requests
pypdf
python-dotenv
httpx
pydantic
numpy
orjson
msgpack
